requirements.txt
fastapi
uvicorn
openai
Pillow
python-multipart
